Refactored the powerup's audio key to a component
