+++ src/entities/Powerup.cpp
@@ -22,6 +22,7 @@ THE SOFTWARE.
 
 #include "Powerup.hpp"
 
+#include "components//Audio.hpp"
 #include "components/AnimatedSprite.hpp"
 #include "components/Lifetime.hpp"
 #include "components/Position.hpp"
@@ -50,7 +51,6 @@ namespace entities
         // Powerups know how to read themselves from the configuration
         std::string imageKey = "image/powerup-" + key;
         auto texture = Content::get<sf::Texture>(imageKey);
-        m_audioKey = "audio/powerup-" + key;
 
         auto spriteCount = Configuration::get<std::uint8_t>(POWERUP_SPRITE_COUNT);
         auto lifetime = misc::msTous(Configuration::get<std::chrono::milliseconds>(POWERUP_LIFETIME));
@@ -61,6 +61,7 @@ namespace entities
         this->addComponent(std::make_unique<components::Size>(math::Dimension2f(size, size)));
         this->addComponent(std::make_unique<components::Lifetime>(lifetime));
         this->addComponent(std::make_unique<components::AnimatedSprite>(texture, spriteCount, spriteTime));
+        this->addComponent(std::make_unique<components::Audio>("audio/powerup-" + key));
 
         // Have to adjust the width dimension by the number of sprites in the image in
         // order for the rendering size to come out correctly.
